Moved activities list to a separate docutment and replaced with itenirary
My position is that we shouldn't give the the choice of activites,
but rather select a complementary set. This has the benefit of:
* reducing time wasted between sessions
* reducing decision exhaustion on the part of the attendees
* giving us, the facilitators, the opportunity to be fully versed in the activites
* gives us the ability to have printouts ready
* allows us to think of prompt questions for the retrospective

I would hope that attendees may want to continues katas by themselves
after the event an in that case a good reference to other katas in a
separate doc would help them a lot.